user profile service

+++ src/main/java/kr/co/pincoin/api/domain/auth/repository/profile/ProfileRepository.java
@@ -8,7 +8,7 @@ import java.util.Optional;
 public interface ProfileRepository {
     Profile save(Profile profile);
 
-    Optional<Profile> findByUserIdWithFetch(Long userId);
+    Optional<Profile> findByUserIdWithFetch(Integer userId);
 
     Optional<Profile> findByUserWithFetch(User user);
