+++ src/main/java/kr/co/pincoin/api/infra/auth/repository/profile/ProfileJpaRepository.java
@@ -12,8 +12,8 @@ import java.util.Optional;
 @Repository
 public interface ProfileJpaRepository extends JpaRepository<ProfileEntity, Long> {
     @Query("SELECT p FROM ProfileEntity p JOIN FETCH p.user u WHERE u.id = :userId")
-    Optional<ProfileEntity> findByUserIdWithFetch(@Param("userId") Long userId);
-    
+    Optional<ProfileEntity> findByUserIdWithFetch(@Param("userId") Integer userId);
+
     @Query("SELECT p FROM ProfileEntity p JOIN FETCH p.user WHERE p.user = :user")
     Optional<ProfileEntity> findByUserWithFetch(@Param("user") User user);
 }
